plugin/full-logger: Extract decrypt into shared lib
